+++ tests/testthat/test_to_tune.R
@@ -396,5 +396,5 @@ test_that("logscale in tunetoken", {
   expect_output(print(to_tune(lower = 1, logscale = TRUE)), "range \\[1, \\.\\.\\.] \\(log scale\\)")
   expect_output(print(to_tune(upper = 1, logscale = TRUE)), "range \\[\\.\\.\\., 1] \\(log scale\\)")
   expect_output(print(to_tune(lower = 0, upper = 1, logscale = TRUE)), "range \\[0, 1] \\(log scale\\)")
-
+  expect_output(print(in_tune()), "Inner")
 })